requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.12.0
pycocotools>=2.0.2
Pillow>=8.3.0
tensorboard>=2.6.0
numpy>=1.21.0